requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
Django==1.8
django-celery==3.1.16
django-cors-middleware==1.2.0
django-filter==0.13.0
django-kombu==0.9.4
django-redis-cache==0.13.1
djangorestframework==3.3.3
kombu==3.0.35
Markdown==2.6.6
Pillow==3.2.0
pytz==2016.4
redis==2.10.5
redis-simple-cache==0.0.8
